backend/requirements-auth.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
numpy
python-dotenv==1.0.0
PyJWT==2.8.0
phonenumbers==8.13.23